backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
tensorflow-cpu==2.15.0
